fix: 大屏恢复
