+++ config/routes.ts
@@ -465,54 +465,54 @@ export const routes = [
             tenant: ['admin'],
             iconfont: 'icon-icon-',
             version: 'pro',
-            authority: ['big-screen'],
-            // authority: ['big-screen', 'vis-configuration'],
+            // authority: ['big-screen'],
+            authority: ['big-screen', 'vis-configuration'],
             routes: [
+              // {
+              //     path: '/data-screen/big-screen',
+              //     name: '大屏管理',
+              //     icon: 'fund',
+              //     tenant: ['admin', 'big-screen'],
+              //     iconfont: 'icon-screen',
+              //     authority: ['big-screen'],
+              //     version: 'pro',
+              //     component: './data-screen/screen',
+              // }
               {
-                  path: '/data-screen/big-screen',
-                  name: '大屏管理',
-                  icon: 'fund',
-                  tenant: ['admin', 'big-screen'],
-                  iconfont: 'icon-screen',
-                  authority: ['big-screen'],
-                  version: 'pro',
-                  component: './data-screen/screen',
+                path: '/data-screen/category',
+                name: '分类管理',
+                icon: 'appstore',
+                tenant: ['admin'],
+                iconfont: 'icon-category-search-fill',
+                authority: ['big-screen'],
+                version: 'pro',
+                component: './data-screen/category',
+              },
+              {
+                path: '/data-screen/screen',
+                name: '大屏管理',
+                icon: 'fund',
+                tenant: ['admin', 'big-screen'],
+                iconfont: 'icon-screen',
+                authority: ['big-screen'],
+                version: 'pro',
+                component: './data-screen/screen',
+              },
+              {
+                path: '/data-screen/configuration',
+                name: '组态管理',
+                icon: 'fund',
+                iconfont: 'icon-screen',
+                authority: ['vis-configuration'],
+                version: 'pro',
+                component: './data-screen/visConfiguration',
+              },
+              {
+                path: '/data-screen/node',
+                name: '组件管理',
+                icon: 'fund',
+                component: './data-screen/node',
               }
-              // {
-              //   path: '/data-screen/category',
-              //   name: '分类管理',
-              //   icon: 'appstore',
-              //   tenant: ['admin'],
-              //   iconfont: 'icon-category-search-fill',
-              //   authority: ['big-screen'],
-              //   version: 'pro',
-              //   component: './data-screen/category',
-              // },
-              // {
-              //   path: '/data-screen/screen',
-              //   name: '大屏管理',
-              //   icon: 'fund',
-              //   tenant: ['admin', 'big-screen'],
-              //   iconfont: 'icon-screen',
-              //   authority: ['big-screen'],
-              //   version: 'pro',
-              //   component: './data-screen/screen',
-              // },
-              // {
-              //   path: '/data-screen/configuration',
-              //   name: '组态管理',
-              //   icon: 'fund',
-              //   iconfont: 'icon-screen',
-              //   authority: ['vis-configuration'],
-              //   version: 'pro',
-              //   component: './data-screen/visConfiguration',
-              // },
-              // {
-              //   path: '/data-screen/node',
-              //   name: '组件管理',
-              //   icon: 'fund',
-              //   component: './data-screen/node',
-              // },
             ],
           },
           {
